Adds issue assignee functionality (#12)
Enables users to assign or unassign issues to project members.

This commit introduces a new command `change_issue_assignee` to update the assignee of an issue.
It also includes UI changes to select the assignee in the issue detail sheet.

The UI allows users to select an assignee from a dropdown list of project members or unassign the issue.

Co-authored-by: Oscar Garcia de Lara <github@ogdlp.mozmail.com>